LayoutTests/fast/dom/background-shorthand-csstext-expected.txt

This page tests whether or not the background shorthand properly omits initial values.

PASS: document.body.style.background should be green and is.
PASS: document.getElementById('div1').style.background should be and is.
PASS: document.getElementById('div2').style.background should be 50% 50% blue and is.
FAIL: document.getElementById('div3').style.background should be repeat rgb(255, 255, 255) but instead is none repeat scroll rgb(255, 255, 255).
